+++ package.json
@@ -48,5 +48,5 @@
     "prettier": "^1.19.1",
     "prettier-eslint": "^9.0.1",
     "rimraf": "^3.0.2"
-  },
+  }
 }
