+++ packages/dockest/package.json
@@ -56,6 +56,7 @@
     "jest": "^29.5.0",
     "mockdate": "^2.0.5",
     "ts-jest": "^29.1.0",
+    "ts-node": "^10.9.1",
     "typescript": "^5.1.3"
   }
 }
