+++ packages/dockest/src/readiness-check/zeroExitCodeReadinessCheck.ts
@@ -1,7 +1,7 @@
 import { race } from 'rxjs'
 import { filter, map, mapTo, take } from 'rxjs/operators'
 import { ReadinessCheck } from '../@types'
-import { DockestError } from '../Errors'
+import { DockestError } from '../_errors'
 import { isDieEvent, isKillEvent } from '../run/bootstrap/createDockerEventEmitter'
 
 /**
