+++ packages/dockest/src/run/waitForServices/index.ts
@@ -1,17 +1,17 @@
 import toposort from 'toposort'
 import { checkConnection } from './checkConnection'
-import { runReadinessCheck } from './runReadinessCheck'
 import { dockerComposeUp } from './dockerComposeUp'
 import { fixRunnerHostAccessOnLinux } from './fixRunnerHostAccessOnLinux'
 import { resolveContainerId } from './resolveContainerId'
+import { runReadinessCheck } from './runReadinessCheck'
 import { runRunnerCommands } from './runRunnerCommands'
-import { createBridgeNetwork } from '../../utils/network/createBridgeNetwork'
-import { DOCKEST_HOST_ADDRESS } from '../../constants'
 import { DockestConfig, Runner } from '../../@types'
-import { joinBridgeNetwork } from '../../utils/network/joinBridgeNetwork'
+import { DockestError } from '../../_errors'
+import { DOCKEST_HOST_ADDRESS } from '../../constants'
 import { bridgeNetworkExists } from '../../utils/network/bridgeNetworkExists'
+import { createBridgeNetwork } from '../../utils/network/createBridgeNetwork'
+import { joinBridgeNetwork } from '../../utils/network/joinBridgeNetwork'
 import { LogWriter } from '../log-writer'
-import { DockestError } from '../../Errors'
 
 const LOG_PREFIX = '[Setup]'
